added authentication and authorisation

+++ api/views.py
@@ -1,58 +1,75 @@
-from django.shortcuts import redirect, render
 from rest_framework import status
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from .serializers import *
 from .models import *
+from django.contrib.auth import authenticate, login as auth_login, logout as auth_logout
+from django.shortcuts import render, redirect
+from django.contrib.auth.models import User
 
 
 def home(request):
-    user_id = request.GET.get('user')
-    user = User.objects.get(id=user_id) if user_id else None
-    user_email = (user.email).split('@')[0]
-    return render(request, 'api/todo.html', {'user': user, 'user_email': user_email})
+    if request.user.is_authenticated:
+        user_id = request.GET.get('user')
+        user = User.objects.get(id=user_id) if user_id else None
+        user_email = (user.email).split('@')[0]
+        return render(request, 'api/todo.html', {'user': user, 'user_email': user_email})
+    else:
+        return redirect('login')
 
 
 def loginpage(request):
     return render(request, 'api/login.html')
 
 
-def login(request):
+def user_login(request):
     if request.method == 'POST':
         email = request.POST['login-email']
         password = request.POST['login-password']
-        try:
-            user = User.objects.get(email=email, password=password)
-            if user:
-                redirect_url = f'/todo-list?user={user.id}'
-                return redirect(redirect_url)
-            else:
-                return render(request, 'api/login.html', {'error': 'Invalid credentials'})
-        except User.DoesNotExist:
+        user = authenticate(request, username=email, password=password)
+        print(user)
+        if user is not None:
+            auth_login(request, user)
+            request.session['email'] = email
+            request.session.save()
+            redirect_url = f'/todo-list?user={user.id}'
+            return redirect(redirect_url)
+        else:
             return render(request, 'api/login.html', {'error': 'Invalid credentials'})
     return render(request, 'api/login.html')
+
+
+def logout(request):
+    auth_logout(request)
+    try:
+        del request.session['email']
+    except KeyError:
+        pass
+    return redirect('login')
 
 
 def signup(request):
     if request.method == 'POST':
         email = request.POST['signup-email']
         password = request.POST['signup-password']
-        user = User(email=email, password=password)
-        user.save()
-        return render(request, 'api/login.html')
+        user = User.objects.create_user(
+            username=email, email=email, password=password)
+        return redirect('login')
     return render(request, 'api/login.html')
 
 
 @api_view(['POST'])
 def add_task(request):
-    print(request.data)
     try:
-        serializer = TaskSerializer(data=request.data, partial=True)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        if request.user.is_authenticated:
+            serializer = TaskSerializer(data=request.data, partial=True)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            else:
+                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         else:
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return Response({"message": "User not authenticated"}, status=status.HTTP_401_UNAUTHORIZED)
     except Exception as e:
         return Response({'error': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
@@ -60,9 +77,12 @@ def add_task(request):
 @api_view(['PUT'])
 def delete_task(request, id):
     try:
-        task = Todo.objects.get(id=id)
-        task.delete()
-        return Response({"message": "Task deleted successfully"}, status=status.HTTP_200_OK)
+        if request.user.is_authenticated:
+            task = Todo.objects.get(id=id)
+            task.delete()
+            return Response({"message": "Task deleted successfully"}, status=status.HTTP_200_OK)
+        else:
+            return Response({"message": "User not authenticated"}, status=status.HTTP_401_UNAUTHORIZED)
     except KeyError:
         return Response({"message": "Task ID is required"}, status=status.HTTP_400_BAD_REQUEST)
     except Todo.DoesNotExist:
@@ -72,17 +92,20 @@ def delete_task(request, id):
 @api_view(['GET'])
 def display_tasks(request):
     try:
-        user_id = request.GET.get('user_id')
-        if user_id:
-            tasks = Todo.objects.filter(user=user_id)
+        if request.user.is_authenticated:
+            user_id = request.GET.get('user_id')
+            if user_id:
+                tasks = Todo.objects.filter(user=user_id)
+            else:
+                return Response({"message": "User ID is required"}, status=status.HTTP_400_BAD_REQUEST)
+            tasks_serializer = AllTaskSerializer(
+                tasks, many=True, context={'request': request})
+            data = {
+                'tasks': tasks_serializer.data
+            }
+            return Response(data, status=status.HTTP_200_OK)
         else:
-            return Response({"message": "User ID is required"}, status=status.HTTP_400_BAD_REQUEST)
-        tasks_serializer = AllTaskSerializer(
-            tasks, many=True, context={'request': request})
-        data = {
-            'tasks': tasks_serializer.data
-        }
-        return Response(data, status=status.HTTP_200_OK)
+            return Response({"message": "User not authenticated"}, status=status.HTTP_401_UNAUTHORIZED)
     except Exception as e:
         return Response({'error': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
@@ -90,7 +113,10 @@ def display_tasks(request):
 @api_view(['PUT'])
 def toggle_task(request, id):
     try:
-        task = Todo.objects.get(id=id)
+        if request.user.is_authenticated:
+            task = Todo.objects.get(id=id)
+        else:
+            return Response({"message": "User not authenticated"}, status=status.HTTP_401_UNAUTHORIZED)
     except Todo.DoesNotExist:
         return Response({"message": "Task not found"}, status=status.HTTP_404_NOT_FOUND)
